Also move advanced network settings to a separate view
